requirements.txt
streamlit
PyGithub
requests